requirements.txt
imutils==0.5.4
numpy==1.24.1
opencv_python==4.7.0.68
Pillow==10.0.0
pytesseract==0.3.10
scikit_image==0.19.3
scikit_learn==1.2.0
ttkthemes==3.2.2